requirements.txt
asgiref==3.2.10
Django==3.0.5
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djongo==1.3.3
pymongo==3.10.1
pytz==2020.1
sqlparse==0.2.4
